feat: hit_circle() returns minimal t value of intersection
hit_circle() returns the smallest value of t such that there is an
intersection with the circle. This is then used by the get_color()
function in order to colour the sphere based on the surface normals.
Normals can be easily calculated with vector geometry. The components
are then rescaled to be between [0,1] and then mapped to the RGB values
respectively.